Homogenized both test to use promise assertions
